Make rlwrap artefacting fix optional
The earlier "rlwrap artefact" fix/work-around would actually cause some
REPLs such as `lein repl` to artefact too (although not as badly).

This change makes the usage of the rlwrap fix optional with the `rlwrap`
key in `repl_config` and will be auto-enabled when the REPL command is
prefixed with "rlwrap".

Related: #9, #15
Co-authored-by: Paul Fernandez <paul4nandez@gmail.com>